Create largest prime factor
functions to find the largest prime number. However, I can't seem to figure out how to identify which numbers are prime numbers, and I can't seem to figure out how to make the algorithm run efficiently. At its current state, it runs the entire range of 600851475143 under the for loop, therefore the algorithm takes a lot of time to run.