Create smallest multiple
The answer is: 232792560

This question was hard until I realized that half of our work is done for us because the prompt tells us that 2520 is the smallest multiples of 1 to 10. If we search for numbers that aren't the product of 2520, they wouldn't be divisible by 1-20.  So my number search is narrowed down to products of 2520. I created a function that checks if a list of divisors are multiples to inputted number. The number would increment by 2520 until the function multiplecheck passes as True.