Time: 8:20 pm to 9:30 pm
I am quite proud of this exercise. I had to draw skills from lab 2 to make the sliding window that checks the product of 13 adjacent numbers by one increment. Didn't realize that we cant use int in for loops, so I cheated and have my function parameters ask for a str of the number instead. 

Adjacent number: 5576689664895
Max product: 23514624000